backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.8.0
flask-mail==0.10.0
pymongo==4.8.0
bcrypt==4.1.3
PyJWT==2.9.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0